requirements.txt
discord.py==1.3.4
Wand==0.6.2
